# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type LeadStatus = typeof LeadStatus[keyof typeof LeadStatus];


export const LeadStatus = {
  new: 'new',
  contacted: 'contacted',
  qualified: 'qualified',
  closed: 'closed',
} as const;

export interface Lead {
  id: number;
  name: string;
  email: string;
  message: string;
  status: LeadStatus;
  createdAt: string;
  updatedAt: string;
}

export interface LeadInput {
  /**
     * @minLength 2
     * @maxLength 120
     */
  name: string;
  /** @maxLength 255 */
  email: string;
  /**
     * @minLength 10
     * @maxLength 4000
     */
  message: string;
}

export interface LeadUpdate {
  status: LeadStatus;
}

export interface SiteSettings {
  heroTitle: string;
  heroAccent: string;
  heroDescription: string;
  monthlyPrice: number;
  pricingBadge: string;
  contactEmail: string;
  updatedAt: string;
}

export interface SiteSettingsInput {
  /**
     * @minLength 3
     * @maxLength 80
     */
  heroTitle: string;
  /**
     * @minLength 3
     * @maxLength 80
     */
  heroAccent: string;
  /**
     * @minLength 10
     * @maxLength 300
     */
  heroDescription: string;
  /**
     * @minimum 0
     * @maximum 10000
     */
  monthlyPrice: number;
  /**
     * @minLength 2
     * @maxLength 80
     */
  pricingBadge: string;
  /** @maxLength 255 */
  contactEmail: string;
}

export type AnalyticsEventType = typeof AnalyticsEventType[keyof typeof AnalyticsEventType];


export const AnalyticsEventType = {
  page_view: 'page_view',
  cta_click: 'cta_click',
  contact_submission: 'contact_submission',
} as const;

export interface AnalyticsEventInput {
  type: AnalyticsEventType;
  /** @maxLength 120 */
  label?: string;
}

export interface AnalyticsDay {
  date: string;
  pageViews: number;
  ctaClicks: number;
  contactSubmissions: number;
}

export interface AnalyticsSummary {
  days: number;
  pageViews: number;
  ctaClicks: number;
  contactSubmissions: number;
  conversionRate: number;
  series: AnalyticsDay[];
}

export interface AdminIdentity {
  email: string;
  isAdmin: boolean;
  allowlistConfigured: boolean;
}

export interface AdminOverview {
  totalLeads: number;
  newLeads: number;
  qualifiedLeads: number;
  pageViews: number;
  conversionRate: number;
  recentLeads: Lead[];
}

export interface BillingStatus {
  provider: string;
  connected: boolean;
  monthlyPrice: number;
  planName: string;
  message: string;
}

export type BillingPriceInterval = typeof BillingPriceInterval[keyof typeof BillingPriceInterval];


export const BillingPriceInterval = {
  month: 'month',
  year: 'year',
} as const;

export interface BillingPrice {
  id: string;
  amount: number;
  currency: string;
  interval: BillingPriceInterval;
}

export interface CheckoutInput {
  /**
     * @minLength 3
     * @maxLength 120
     */
  priceId: string;
}

export interface CheckoutSession {
  url: string;
}

export interface BillingPortalInput {
  /**
     * @minLength 3
     * @maxLength 200
     */
  sessionId: string;
}

export type ListAdminLeadsParams = {
search?: string;
status?: LeadStatus;
};

export type GetAdminAnalyticsParams = {
days?: GetAdminAnalyticsDays;
};

export type GetAdminAnalyticsDays = typeof GetAdminAnalyticsDays[keyof typeof GetAdminAnalyticsDays];


export const GetAdminAnalyticsDays = {
  NUMBER_7: 7,
  NUMBER_30: 30,
  NUMBER_90: 90,
} as const;