# lib/api-zod/src/generated/types/billingPortalInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface BillingPortalInput {
  /**
     * @minLength 3
     * @maxLength 200
     */
  sessionId: string;
}