# lib/api-zod/src/generated/types/billingPrice.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BillingPriceInterval } from './billingPriceInterval';

export interface BillingPrice {
  id: string;
  amount: number;
  currency: string;
  interval: BillingPriceInterval;
}