# lib/api-zod/src/generated/types/checkoutInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CheckoutInput {
  /**
     * @minLength 3
     * @maxLength 120
     */
  priceId: string;
}